Moved the unit testing project into its own folder
