+++ Test/Fido2.Tests/Fido2.Tests.csproj
@@ -15,9 +15,9 @@
   </ItemGroup>
 
   <ItemGroup>
-    <ProjectReference Include="..\Src\Fido2.AspNet\Fido2.AspNet.csproj" />
-    <ProjectReference Include="..\Src\Fido2.Models\Fido2.Models.csproj" />
-    <ProjectReference Include="..\Src\Fido2\Fido2.csproj" />
+    <ProjectReference Include="..\..\Src\Fido2.AspNet\Fido2.AspNet.csproj" />
+    <ProjectReference Include="..\..\Src\Fido2.Models\Fido2.Models.csproj" />
+    <ProjectReference Include="..\..\Src\Fido2\Fido2.csproj" />
     <PackageReference Include="Microsoft.Extensions.Caching.Memory" Version="2.2.0" />
     <PackageReference Include="Microsoft.Extensions.DependencyInjection" Version="2.2.0" />
     <PackageReference Include="Microsoft.Extensions.Logging" Version="2.2.0" />
