+++ Test/Fido2.Tests/TestFiles/attestationNoneOptions.json
