+++ Test/Fido2.Tests/TestFiles/attestationOptionsNone.json
