+++ Test/Fido2.Tests/TestFiles/attestationOptionsPacked.json
