+++ Test/Fido2.Tests/TestFiles/attestationOptionsPacked512.json
