+++ Test/Fido2.Tests/TestFiles/attestationResultsPacked512.json
