+++ Test/Fido2.Tests/TestFiles/attestationTPMSHA1Options.json
